WIP: Pre-download source files to avoid racy truncations

+++ gfw_pixetl/utils/utils.py
@@ -136,7 +136,7 @@ def available_memory_per_process_bytes() -> float:
 
 
 def available_memory_per_process_mb() -> float:
-    mem = GLOBALS.max_mem / GLOBALS.workers
+    mem = int(GLOBALS.max_mem / GLOBALS.workers)
     LOGGER.info(f"Available memory per worker set to {mem} MB")
     return mem
